rename: hooks/api -> hooks/query 및 폴더 공통화
