+++ src/components/Main/PersonalFeedList.tsx
@@ -3,8 +3,8 @@ import { useMemo } from 'react';
 
 import { Title } from '@/components/common';
 import { Feed } from '@/components/Travelogue/Feed';
-import { useGetPersonalTravelogues } from '@/hooks/api/travelogue';
 import useIntersect from '@/hooks/common/useIntersect';
+import { useGetPersonalTravelogues } from '@/hooks/query/main';
 import { flexCenterStyle } from '@/styles/commonStyle';
 
 const PersonalFeedList = () => {
