+++ src/hooks/query/detail.ts
@@ -1,0 +1,13 @@
+import { useMutation } from '@tanstack/react-query';
+
+import { patchTravelogueDetailById } from '@/api/main';
+
+export const usePatchTravelogueDetailById = () => {
+  return useMutation({
+    mutationFn: async (data: { travelogueId: string }) =>
+      await patchTravelogueDetailById(data),
+    onError: (error: { message: string }) => {
+      console.error(error.message);
+    },
+  });
+};
