+++ src/hooks/query/search.ts
@@ -1,0 +1,55 @@
+import {
+  QueryFunctionContext,
+  useInfiniteQuery,
+  UseInfiniteQueryResult,
+} from '@tanstack/react-query';
+import { AxiosResponse } from 'axios';
+
+import { getTravelogueListByFilter } from '@/api/search';
+import { FilterAxiosProps } from '@/types/filter';
+import { TravelogueListType } from '@/types/travelogue';
+
+export const useGetTravelogueByFilter = ({
+  keyword,
+  size,
+  minDays,
+  maxDays,
+  minCost,
+  maxCost,
+  sort,
+}: FilterAxiosProps): UseInfiniteQueryResult<
+  AxiosResponse<TravelogueListType>,
+  unknown
+> => {
+  return useInfiniteQuery(
+    ['FILTER_TRAVELOGUES', keyword, minDays, maxDays, minCost, maxCost, sort],
+    ({ pageParam = 0 }: QueryFunctionContext) =>
+      getTravelogueListByFilter({
+        keyword,
+        size,
+        page: pageParam,
+        minDays,
+        maxDays,
+        minCost,
+        maxCost,
+        sort,
+      }),
+    {
+      getNextPageParam: ({ data: { last, number } }) => (last ? undefined : number + 1),
+    },
+  );
+};
+
+// export const useGetTravelogueByKeyword = (
+//   keyword: string,
+//   size: number,
+// ): UseInfiniteQueryResult<AxiosResponse<TravelogueListType>, unknown> => {
+//   return useInfiniteQuery(
+//     ['KEYWORD_TRAVELOGUES', keyword],
+//     ({ pageParam = 0 }: QueryFunctionContext) =>
+//       getTravelogueListByKeyword(keyword, pageParam, size),
+//     {
+//       getNextPageParam: ({ data: { last, number } }) => (last ? undefined : number + 1),
+//     },
+//   );
+// };
